+++ server/src/server.ts
@@ -7,6 +7,7 @@ import {PORT} from './configs/environment';
 import logger from './configs/winston';
 import {connectToDatabase} from './database/index';
 
+// eslint-disable-next-line @typescript-eslint/no-misused-promises
 app.listen(PORT, async () => {
   try {
     await connectToDatabase();
